feat: swap and landing page added
