+++ lib/pages/ExplorePage.dart
@@ -1,0 +1,79 @@
+import 'package:ae_dapp/pages/NavigationBar.dart';
+import 'package:flutter/material.dart';
+
+class ExplorePage extends StatefulWidget {
+  const ExplorePage({Key? key}) : super(key: key);
+
+  @override
+  _ExplorePageState createState() => _ExplorePageState();
+}
+
+class _ExplorePageState extends State<ExplorePage> {
+
+  double lgTxSize = 52;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Stack(
+        children: <Widget>[
+          Image(
+            image: AssetImage("assets/images/background.jpeg"),
+            width: MediaQuery.of(context).size.width,
+            height: MediaQuery.of(context).size.height,
+            fit: BoxFit.cover,
+          ),
+          Column(
+            children: <Widget>[
+              Align(
+                alignment: Alignment.topRight,
+                child: Padding(
+                  padding: const EdgeInsets.only(right: 20),
+                  child: Image(
+                    width: 80,
+                    height: 80,
+                    image: AssetImage("assets/images/x.png"),
+                  ),
+                )
+              ),
+              Text(
+                "EXPLORE",
+                style: TextStyle(
+                  fontFamily: 'OpenSans',
+                  fontSize: lgTxSize,
+                  fontWeight: FontWeight.w600,
+                  fontStyle: FontStyle.italic,
+                )
+              ),
+              Container(
+                color: Colors.grey[900],
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: <Widget>[
+                    Column(
+                      children: <Widget>[
+                        Row(
+                          children: [
+                            Text(
+                              "Athlete Tokens"
+                            ),
+                          ],
+                        )
+                        
+                      ]
+                    ),
+                    Column(
+                      children: <Widget>[
+
+                      ]
+                    ),
+                  ],
+                )
+              )
+            ]
+          )
+        ],
+      )
+    );
+  }
+}
