+++ lib/pages/NavigationBar.dart
@@ -1,9 +1,8 @@
-import 'package:ae_dapp/pages/AthletesList.dart';
+import 'package:ae_dapp/service/AthletesList.dart';
 import 'package:flutter/material.dart';
 import "package:velocity_x/velocity_x.dart";
-import 'package:ae_dapp/pages/wallet.dart';
-import 'package:ae_dapp/pages/MyTeam.dart';
-import 'package:ae_dapp/pages/swapPage.dart';
+import 'package:ae_dapp/pages/MyTeam2.dart';
+import 'package:ae_dapp/pages/TradingBlock.dart';
 import 'package:ae_dapp/pages/AXPage.dart';
 
 class NavigationBar extends StatefulWidget {
@@ -21,10 +20,9 @@ class _NavigationBarState extends State<NavigationBar> {
   // This is where the pages are linked to the navigation
   List<Widget> _widgetOptions = <Widget>[
     AXPage(),
-    Wallet(),
-    MyTeam(),
     AthletesList(),
-    Swap()
+    TradingBlock(),
+    MyTeam2(),
   ];
 
   void _onItemTapped(int index) {
@@ -42,24 +40,24 @@ class _NavigationBarState extends State<NavigationBar> {
       bottomNavigationBar: BottomNavigationBar(
         items: const <BottomNavigationBarItem>[
           BottomNavigationBarItem(
-            icon: Icon(Icons.account_balance_wallet),
-            label: 'Staking',
+            icon: Icon(Icons.attach_money),
+            label: 'AX',
           ),
           // BottomNavigationBarItem(  // Move this into same tab as Buy Athletes
           //   icon: Icon(Icons.account_balance_wallet),
           //   label: 'My Athletes',
           // ),
           BottomNavigationBarItem(
-            icon: Icon(Icons.sports_basketball),
-            label: 'My Team',
-          ),
-          BottomNavigationBarItem(
             icon: Icon(Icons.search),
-            label: 'Search',
+            label: 'Explore',
           ),
           BottomNavigationBarItem(
             icon: Icon(Icons.swap_horizontal_circle_sharp),
-            label: 'Swap',
+            label: 'Trading Block',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(IconData(58094, fontFamily: 'MaterialIcons')),
+            label: 'My Team',
           ),
           // BottomNavigationBarItem(
           //   icon: Icon(Icons.settings),
