+++ lib/service/Portfolio.dart
@@ -1,4 +1,3 @@
-import 'package:ae_dapp/pages/wallet.dart';
 import 'package:chart_sparkline/chart_sparkline.dart';
 import 'package:flutter/material.dart';
