+++ pubspec.yaml
@@ -80,11 +80,10 @@ flutter:
         - asset: assets/fonts/PublicSans-Thin.ttf
 
   assets:
-    - assets/images/img1.png
-    - assets/images/img2.png
+    - assets/images/trophy.png
+    - assets/images/AXMarkets.png
     - assets/images/asu-hero.jpg
-    - assets/images/ax.jpg
-    - assets/images/x.png
+    - assets/images/x.jpg
     - assets/images/x.png
     - assets/images/background.jpeg
